9742: moving trialInfo-setting logic one level firther so both helpers receive messages with trial information on them
