+++ web-client/src/views/WorkQueue/RecentMessagesInbox.jsx
@@ -76,7 +76,15 @@ export const RecentMessagesInbox = connect(
                   <td className="message-queue-row">
                     <span>{item.caseTitle}</span>
                   </td>
-                  <td>{item.caseStatus}</td>
+                  {!item.showTrialInformation && (
+                    <td className="message-queue-row">{item.caseStatus}</td>
+                  )}
+                  {item.showTrialInformation && (
+                    <td className="message-queue-row">
+                      {item.caseStatus} - {item.formattedTrialDate}{' '}
+                      {item.formattedTrialLocation}
+                    </td>
+                  )}
                   <td>{item.from}</td>
                   <td>{item.fromSection}</td>
                 </tr>
